samples/python/24.bot-authentication-msgraph/requirements.txt
botbuilder-integration-aiohttp>=4.13.0
botbuilder-dialogs>=4.13.0
requests_oauthlib

